feat: Add socket parameter to ChatDrawer component
The ChatDrawer component in ChatDrawer.js was modified to include a new socket parameter. This allows the component to interact with the socket connection and send messages. The socket parameter is passed from the parent component and used to emit a "message2" event with the message data when sending a message.